trans2

+++ Application/Interfaces/ITransactionRepository.cs
@@ -2,6 +2,6 @@ namespace Application.Interfaces;
 
 public interface ITransactionRepository
 {
-    Task<int> SaveNewTransaction(int accountId, string bankaccount, DateTime transactionDate, double amount,
+    Task<int> SaveNewTransaction(int accountId, string bankaccount, DateTime transactionDate, Decimal amount,
         int subCategoryId, string transactionDescription);
 }
